+++ Stutter.h
@@ -11,7 +11,7 @@ private:
     double mPhase;
 public:
     void setSampleRate(double sampleRate);
-    void setSpeed(double speed);
+    void setSpeed(int speed);
     void setBPM(double bpm);
     void processSamples(double inputbuffer1, double inputbuffer2, double &outputbuffer1, double &outputbuffer2, int nFrames);
 };
